+++ CHANGELOG.md
@@ -1,7 +1,9 @@
 Release Notes
 =============
 
-## Next Version
+## 2.10.0
+
+Released March 30, 2018 &bull; [diff](https://github.com/groue/GRDB.swift/compare/v2.9.0...v2.10.0)
 
 ### New